Added split_image method and fixed create_tasks_from_file method

+++ slaves.py
@@ -4,7 +4,7 @@ import os
 import cv2 # Lib for image processing
 
 client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-SERVER_HOST = "127.0.1.1" # socket.gethostbyname(socket.gethostname())
+SERVER_HOST = "192.168.1.29" # socket.gethostbyname(socket.gethostname())
 SERVER_PORT = 5051
 
 # Connect to the server
@@ -25,7 +25,6 @@ with open(file_name, "wb") as file:
     received_bytes = 0
     while received_bytes < file_size:
         data = client_socket.recv(1024)
-        print("data :", data)
         file.write(data)
         received_bytes += len(data)
         progress.update(len(data))
@@ -33,33 +32,40 @@ with open(file_name, "wb") as file:
 
 print("[+] Image received successfully.")
 
-# Send "DONE" command
-client_socket.send("DONE".encode())
-print("[*] Sent DONE command to server.")
+# Check received image
+if not os.path.exists(file_name) or os.path.getsize(file_name)==0:
 
-# Process image
-image = cv2.imread(file_name,cv2.IMREAD_COLOR)
-image = cv2.cvtColor(image,cv2.COLOR_BGR2GRAY)
-edges = cv2.Canny(image,100,200)
-processed_file_path = "processedImage.jpg"
-cv2.imwrite(processed_file_path, edges)
+    print(f"Error: The file {file_name} does not exist or is empty.")
 
-# Send image to server
-file_name = processed_file_path
-file_size = os.path.getsize(file_name)
+else:
+    # Send "DONE" command
+    client_socket.send("DONE".encode())
+    print("[*] Sent DONE command to server.")
 
-# Send file name and size
-client_socket.send(f"{file_name}\n{file_size}".encode())
+    # Process image
+    image = cv2.imread(file_name,cv2.IMREAD_COLOR)
+    print(f"Image shape: {image.shape}")
+    image = cv2.cvtColor(image,cv2.COLOR_BGR2GRAY)
+    edges = cv2.Canny(image,100,200)
+    processed_file_path = "processedImage.jpg"
+    cv2.imwrite(processed_file_path, edges)
 
-# Send file data in chunks
-with open(file_name, "rb") as file:
-    while True:
-        data = file.read(1024)
-        if not data:
-            break
-        client_socket.sendall(data)
+    # Send image to server
+    file_name = processed_file_path
+    file_size = os.path.getsize(file_name)
 
-print("[+] Image sent to server successfully.")
+    # Send file name and size
+    client_socket.send(f"{file_name}\n{file_size}".encode())
+
+    # Send file data in chunks
+    with open(file_name, "rb") as file:
+        while True:
+            data = file.read(1024)
+            if not data:
+                break
+            client_socket.sendall(data)
+
+    print("[+] Image sent to server successfully.")
 
 # Close the connection given server response.
 server_response = client_socket.recv(1024).decode()
